+++ client/src/components/AssignmentBoard/VolunteerModal/content/Activity.jsx
@@ -1,18 +1,13 @@
 /* eslint-disable eqeqeq */
 /* eslint-disable react/destructuring-assignment */
-import React, { useState, useEffect } from 'react';
+import React, { useState } from 'react';
 import { PlusCircle } from 'lucide-react';
 import ActivityInput from './ActivityInput';
 import Button from '../../../Button';
 import { Label3, BodySubText, BodySubLabel } from '../../../../styles/components/Typography';
 import { ProfileInfoContainer } from '../../../../styles/components/VolunteerModal.style';
-import { StyledInput } from '../../../../styles/components/Input.style';
 
-
-const Activity = ({ events, isEditing, volunteerId, trackActivityChange, addNewActivity }) => {
-    const [updatedActivity, setUpdatedActivity] = useState([]);
-
-    return (
+const Activity = ({ events, isEditing, volunteerId, trackActivityChange, addNewActivity }) => (
     <>
         <h3>Activity</h3>
         <BodySubText>Review and update a volunteer&apos;s progress and status.</BodySubText>
@@ -46,7 +41,7 @@ const Activity = ({ events, isEditing, volunteerId, trackActivityChange, addNewA
             )
         }
     </>
-    );
-};
+);
+
 
 export default Activity;
